# doc/dev-guide/apispec/swagger/rpc.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "title": "api/etcdserverpb/rpc.proto",
    "version": "version not set"
  },
  "paths": {
...
    "/v3/schedule/evals/ack": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "operationId": "Schedule_EvalAck",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbScheduleEvalAckRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/etcdserverpbScheduleEvalAckResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
    "/v3/schedule/evals/dequeue": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "summary": "eval 队列开始",
        "operationId": "Schedule_EvalDequeue",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbScheduleEvalDequeueRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/etcdserverpbScheduleEvalDequeueResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
    "/v3/schedule/evals/nack": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "operationId": "Schedule_EvalNack",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/etcdserverpbScheduleEvalNackRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/etcdserverpbScheduleEvalNackResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "etcdserverpbResponseHeader": {
      "type": "object",
      "properties": {
        "cluster_id": {
          "description": "cluster_id is the ID of the cluster which sent the response.",
          "type": "string",
          "format": "uint64"
        },
        "member_id": {
          "description": "member_id is the ID of the member which sent the response.",
          "type": "string",
          "format": "uint64"
        },
        "raft_term": {
          "description": "raft_term is the raft term when the request was applied.",
          "type": "string",
          "format": "uint64"
        },
        "revision": {
          "description": "revision is the key-value store revision when the request was applied.\nFor watch progress responses, the header.revision indicates progress. All future events\nrecieved in this stream are guaranteed to have a higher revision number than the\nheader.revision number.",
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "etcdserverpbScheduleEvalAckRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "job_id": {
          "type": "string"
        },
        "namespace": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "etcdserverpbScheduleEvalAckResponse": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        }
      }
    },
...
    "etcdserverpbScheduleEvalDequeueRequest": {
      "type": "object",
      "title": "=== schedule alloc end ===\n=== schedule eval queue start ===",
      "properties": {
        "namespace": {
          "type": "string"
        },
        "schedulers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "timeout": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "etcdserverpbScheduleEvalDequeueResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/schedulepbEvaluation"
        },
        "header": {
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        },
        "token": {
          "type": "string"
        }
      }
    },
...
    "etcdserverpbScheduleEvalNackRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "job_id": {
          "type": "string"
        },
        "namespace": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "etcdserverpbScheduleEvalNackResponse": {
      "type": "object",
      "properties": {
        "header": {
          "$ref": "#/definitions/etcdserverpbResponseHeader"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        },
        "error": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "schedulepbEvaluation": {
      "type": "object",
      "properties": {
        "blocked_eval": {
          "type": "string"
        },
        "create_time": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string"
        },
        "job_id": {
          "type": "string"
        },
        "namespace": {
          "type": "string"
        },
        "next_eval": {
          "type": "string"
        },
        "node_id": {
          "type": "string"
        },
        "plan_id": {
          "type": "string"
        },
        "previous_eval": {
          "type": "string"
        },
        "priority": {
          "type": "string",
          "format": "uint64"
        },
        "status": {
          "type": "string"
        },
        "status_description": {
          "type": "string"
        },
        "triggered_by": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "update_time": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKey": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKey": []
    }
  ]
}